requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.20.0
requests>=2.28.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.64.0
asyncio-throttle>=1.0.0
fake-useragent>=1.1.0
openai>=1.0.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Web API依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic>=2.0.0

# 人脸识别增强
# insightface>=0.7.3  # 可选：专业人脸识别
# onnxruntime>=1.16.0 # 可选：支持ONNX模型
